# code/docs/swagger.yaml
basePath: /
definitions:
...
  category.Category:
    properties:
      code:
        type: string
      createdAt:
        type: string
      deactivatedAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      parentID:
        type: string
      sortOrder:
        type: integer
      subCategories:
        items:
          $ref: '#/definitions/category.Category'
        type: array
      type:
        $ref: '#/definitions/category.CategoryType'
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  category.CategoryListResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/category.Category'
        type: array
    type: object
  category.CategoryResponse:
    properties:
      category:
        $ref: '#/definitions/category.Category'
    type: object
  category.CategoryType:
    enum:
    - INCOME
    - EXPENSE
    - MOVEMENT
    type: string
    x-enum-comments:
      CategoryTypeMovement: special, can't be created via endpoint
    x-enum-descriptions:
    - ""
    - ""
    - special, can't be created via endpoint
    x-enum-varnames:
    - CategoryTypeIncome
    - CategoryTypeExpense
    - CategoryTypeMovement
  category.ChangeCategoryInput:
    properties:
      description:
        type: string
      name:
        type: string
      parent_code:
        type: string
      type:
        type: string
    type: object
  category.CreateCategoryInput:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_code:
        type: string
      type:
        type: string
    required:
    - code
    - description
    - name
    - type
    type: object
  category.ReorderCategoriesInput:
    properties:
      codes:
        items:
          type: string
        type: array
      parent_code:
        type: string
    required:
    - codes
    type: object
  handlers.ErrorDetail:
    properties:
      code:
        example: invalid_input
        type: string
      error:
        example: invalid request body
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/handlers.ErrorDetail'
    type: object
  reports.AccountBalance:
    properties:
      accountCode:
        type: string
      balance:
        format: int64
        type: integer
    type: object
  reports.AllBalancesResponse:
    properties:
      accounts:
        items:
          $ref: '#/definitions/reports.AccountBalance'
        type: array
    type: object
  reports.BalanceByCodeResponse:
    properties:
      account:
        $ref: '#/definitions/reports.AccountBalance'
    type: object
  reports.CategoryYearlyBalance:
    properties:
      code:
        type: string
      monthly_data:
        items:
          type: integer
        type: array
      subcategories:
        items:
          $ref: '#/definitions/reports.CategoryYearlyBalance'
        type: array
    type: object
  reports.DashboardData:
    properties:
      balances:
        items:
          $ref: '#/definitions/reports.AccountBalance'
        type: array
      month:
        type: integer
      recent_transactions:
        items:
          $ref: '#/definitions/transaction.TransactionResponseItem'
        type: array
      top_expenses:
        items:
          $ref: '#/definitions/transaction.TransactionResponseItem'
        type: array
      year:
        type: integer
      yearly:
        items:
          $ref: '#/definitions/reports.CategoryYearlyBalance'
        type: array
    type: object
  reports.DashboardResponse:
    properties:
      dashboard:
        $ref: '#/definitions/reports.DashboardData'
    type: object
  reports.YearlyReportResponse:
    properties:
      balances:
        items:
          $ref: '#/definitions/reports.CategoryYearlyBalance'
        type: array
    type: object
...
  transaction.TransactionResponseItem:
    properties:
      account_code:
        type: string
      account_transfer:
        type: string
      amount:
        type: integer
      category_code:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      transfer_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for managing accounts, categories, transactions, reports, and suggestions.
  title: Expense Tracker API
  version: "1.0"
paths:
...
  /categories:
    get:
      description: Returns all active categories for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.CategoryListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Creates a new category for the authenticated user
      parameters:
      - description: Category payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/category.CreateCategoryInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/category.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - categories
  /categories/{code}:
    delete:
      description: Deactivates a category
      parameters:
      - description: Category code
        in: path
        name: code
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deactivate category
      tags:
      - categories
    get:
      description: Returns a category by its code
      parameters:
      - description: Category code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get category by code
      tags:
      - categories
    patch:
      consumes:
      - application/json
      description: Updates an existing category
      parameters:
      - description: Category code
        in: path
        name: code
        required: true
        type: string
      - description: Category changes
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/category.ChangeCategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - categories
  /categories/reorder:
    patch:
      consumes:
      - application/json
      description: Reorders categories within a parent group
      parameters:
      - description: Category ordering
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/category.ReorderCategoriesInput'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reorder categories
      tags:
      - categories
  /reports/balance:
    get:
      description: Returns balances for all accounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reports.AllBalancesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List balances
      tags:
      - reports
  /reports/balance/{code}:
    get:
      description: Returns a single account balance
      parameters:
      - description: Account code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reports.BalanceByCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get balance by account code
      tags:
      - reports
  /reports/dashboard:
    get:
      description: Returns dashboard data for the selected month and year
      parameters:
      - description: Year to filter
        in: query
        minimum: 2000
        name: year
        type: integer
      - description: Month to filter
        in: query
        maximum: 12
        minimum: 1
        name: month
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reports.DashboardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get dashboard
      tags:
      - reports
  /reports/yearly:
    get:
      description: Returns all balances for the specified year (default current year)
      parameters:
      - description: Year to filter
        in: query
        minimum: 2000
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reports.YearlyReportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get yearly report
      tags:
      - reports
...
securityDefinitions:
  BearerAuth:
    description: JWT token using the format "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"